Rename utils to align better with the *.*.ts convention
