+++ client-v2/src/app/pages/auth/login/login.component.ts
@@ -9,7 +9,7 @@ import { LoginCredentialsDto } from 'src/app/models/auth.model'
 import { AppState } from 'src/app/store'
 import { authActions } from 'src/app/store/user/user.actions'
 import { userSelectors } from 'src/app/store/user/user.selectors'
-import { getErrorMapUpdates } from '../../../utils/store-helpers'
+import { getErrorMapUpdates } from '../../../utils/store.helpers'
 
 @Component({
     selector: 'app-login',
