+++ client-v2/src/app/pages/settings/account/account.component.ts
@@ -9,7 +9,7 @@ import { betterEmailValidator, matchSibling } from 'src/app/components/molecules
 import { AppState } from 'src/app/store'
 import { accountActions, authActions } from 'src/app/store/user/user.actions'
 import { moveToMacroQueue } from 'src/app/utils'
-import { getErrorMapUpdates, getLoadingUpdates } from '../../../utils/store-helpers'
+import { getErrorMapUpdates, getLoadingUpdates } from '../../../utils/store.helpers'
 
 @Component({
     templateUrl: './account.component.html',
