+++ client-v2/src/app/store/user/account.effects.ts
@@ -5,7 +5,7 @@ import { Store } from '@ngrx/store'
 import { catchError, map, mergeMap, of } from 'rxjs'
 import { HttpServerErrorResponse } from 'src/app/http/types'
 import { UserService } from 'src/app/services/user.service'
-import { getMessageFromHttpError } from 'src/app/utils/store-helpers'
+import { getMessageFromHttpError } from 'src/app/utils/store.helpers'
 import { AppState } from '..'
 import { appActions } from '../app.actions'
 import { accountActions, authActions } from './user.actions'
