+++ client-v2/src/app/store/user/user.model.ts
@@ -1,5 +1,5 @@
 import { User } from 'src/app/models'
-import { PartialRequired } from 'src/app/utils/type-helpers'
+import { PartialRequired } from 'src/app/utils/type.helpers'
 
 export interface UserState {
     me: PartialRequired<User, 'id' | 'username'> | null
