+++ client-v2/src/app/utils/type.helpers.ts
